requirements.txt
Pillow==12.1.0
streamlit==1.41.0
